feat: EatTogatherMembers 구현
- 기존의 List<EatTogahterMember>을 일급 컬랙션으로 구현
